M5: the bare-metal application builds
The Vitis workspace, platform and application are created and dcc_l4.elf
links: 281 kB text, 1.27 MB bss, 554 kB on disk. board_main.c and
platform_standalone.c were the two files that had never seen a compiler, and
both needed a fix, each Cortex-A9 specific:

  - xtime_l.h does not exist in a Cortex-A9 system-device-tree BSP. It is an
    UltraScale+ (A53/R5) header. The same XTime_GetTime and COUNTS_PER_SECOND
    come from xiltimer.h, which includes xtimer_config.h. Also parenthesise
    COUNTS_PER_SECOND before dividing again: it expands to a bare
    XPAR_CPU_CORE_CLOCK_FREQ_HZ/2 and reassociates otherwise.

  - XAxiDma_Busy returns u32, so %d is a -Wformat error under the default
    -Wall -Wextra.

vitis_setup.py now checks for the ELF after building instead of trusting
comp.build(), which prints "build failed" and returns rather than raising --
the first run reported success over a failed compile.

The runbook gains the verified manual IDE sequence. Sources dropped in the
component's src/ are picked up by aux_source_directory with no registration
step, which is the simpler route by hand; import_files puts them in the
component root and registers them through USER_COMPILE_SOURCES instead.

+++ host/board_main.c
@@ -135,10 +135,10 @@ static int encode_one(dcc_kernel_t *k, int width, int height, int qp,
     }
 
     if (dcc_kernel_wait_done(k, 2000000u) != 0) {
-        printf("FAIL: kernel timeout. STATUS=%08lx, MM2S busy=%d, S2MM busy=%d\n",
+        printf("FAIL: kernel timeout. STATUS=%08lx, MM2S busy=%lu, S2MM busy=%lu\n",
                (unsigned long)dcc_mmio_read(k->base, DCC_REG_STATUS),
-               XAxiDma_Busy(&dma, XAXIDMA_DMA_TO_DEVICE),
-               XAxiDma_Busy(&dma, XAXIDMA_DEVICE_TO_DMA));
+               (unsigned long)XAxiDma_Busy(&dma, XAXIDMA_DMA_TO_DEVICE),
+               (unsigned long)XAxiDma_Busy(&dma, XAXIDMA_DEVICE_TO_DMA));
         return -1;
     }
     /* Let S2MM retire the last beats after tlast before reading the buffer. */
